pyproject.toml
[project]
name = "notion-rag"
version = "0.1.0"
description = "RAG service for Notion workspace knowledge base"
requires-python = ">=3.11"
dependencies = [
    # Notion API
    "notion-client>=2.2.0",
    # Vector DB
    "qdrant-client>=1.12.0",
    # Embedding
    "fastembed>=0.4.0",
    # Text processing
    "langchain-text-splitters>=0.3.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Config
    "pydantic-settings>=2.6.0",
    # Logging
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
